SLES15 and Specv8 Improvement
This commit include two changes:
1. Modify code to be more backward compatible, so that python3.6 can work.
As SDE and SLES15 works better with older version of python.
2. Specv8 compiler report flow improvement. Ignore VE error, also
continue to process other workloads even error found in global csv for
one benchmark.

+++ copy_files.py
@@ -19,11 +19,14 @@ if __name__ == '__main__':
         for row in csv_reader:
             new_row = row.copy()
             sim_files = row['sim_files'].split(',')
+            status = row['status']
             exe = row['exe']
             new_row['exe'] = os.path.basename(exe)
             new_row['sim_files'] = ','.join([os.path.basename(sim_file) for sim_file in sim_files])
 
             sub_dir = os.path.join(args.dst, new_row['name'])
+            if status != "Success":
+                 continue
             if os.path.isdir(sub_dir):
                  shutil.rmtree(sub_dir)
             os.mkdir(sub_dir)
